chore: rename Search2 to Search
