+++ src/app/Scenes/Home/Home.tsx
@@ -21,7 +21,7 @@ import { Home_homePageBelow$data } from "__generated__/Home_homePageBelow.graphq
 import { Home_meAbove$data } from "__generated__/Home_meAbove.graphql"
 import { Home_meBelow$data } from "__generated__/Home_meBelow.graphql"
 import { Home_newWorksForYou$data } from "__generated__/Home_newWorksForYou.graphql"
-import { Search2Query } from "__generated__/Search2Query.graphql"
+import { SearchQuery } from "__generated__/SearchQuery.graphql"
 import { AboveTheFoldFlatList } from "app/Components/AboveTheFoldFlatList"
 import { LargeArtworkRailPlaceholder } from "app/Components/ArtworkRail/LargeArtworkRail"
 import { ArtistRailFragmentContainer } from "app/Components/Home/ArtistRails/ArtistRail"
@@ -49,7 +49,7 @@ import {
   DEFAULT_RECS_MODEL_VERSION,
   RECOMMENDATION_MODEL_EXPERIMENT_NAME,
 } from "app/Scenes/NewWorksForYou/NewWorksForYou"
-import { search2QueryDefaultVariables } from "app/Scenes/Search/Search2"
+import { searchQueryDefaultVariables } from "app/Scenes/Search/Search"
 import { ViewingRoomsHomeMainRail } from "app/Scenes/ViewingRoom/Components/ViewingRoomsHomeRail"
 import { GlobalStore } from "app/store/GlobalStore"
 import { getRelayEnvironment } from "app/system/relay/defaultEnvironment"
@@ -144,7 +144,7 @@ const Home = memo((props: HomeProps) => {
   }).current
 
   useEffect(() => {
-    prefetchUrl<Search2Query>("search", search2QueryDefaultVariables)
+    prefetchUrl<SearchQuery>("search", searchQueryDefaultVariables)
     prefetchUrl("my-profile")
     prefetchUrl("inbox")
     prefetchUrl("sales")
